backend/requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
gcloud==0.17.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
jws==0.1.3
oauth2client==3.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
